Move BlockEventSource to BlockEvent as sealed class Source.
