+++ src/main/java/net/minestom/server/event/block/PreBreakBlockEvent.java
@@ -24,14 +24,14 @@ public final class PreBreakBlockEvent implements Event, BlockEvent, CancellableE
     private final Instance instance;
     private final BlockFace blockFace;
     private final BlockVec position;
-    private final BlockEventSource source;
+    private final BlockEvent.Source source;
 
     public PreBreakBlockEvent(
         @NotNull Block previousBlock,
         @NotNull Instance instance,
         @Nullable BlockFace face,
         @NotNull BlockVec position,
-        @NotNull BlockEventSource source
+        @NotNull BlockEvent.Source source
     ) {
         this.previousBlock = previousBlock;
         this.instance = instance;
@@ -88,11 +88,11 @@ public final class PreBreakBlockEvent implements Event, BlockEvent, CancellableE
     }
 
     /**
-     * Gets the {@link BlockEventSource}
+     * Gets the {@link BlockEvent.Source}
      *
-     * @return the BlockEventSource
+     * @return the Events Source
      */
-    public @NotNull BlockEventSource getSource() {
+    public @NotNull BlockEvent.Source getSource() {
         return source;
     }
